Tests/LibWeb/Text/input/hit_testing/empty-anonymous-container-overlapping.html

<!DOCTYPE html>
<style>
#a {
    margin-bottom: -20px;
}
body {
    border: 1px solid blue;
}
</style>
<script src="../include.js"></script>
<div id="a"><a href="#" id="target">You should be able to click this link</a></div><br>
<script>
test(() => {
    const hit = internals.hitTest(50, 18);
    printElement(hit.node);
    println(`index: ${hit.indexInNode}`);
    printElement(hit.node.parentNode);
});
</script>
