Tests/LibWeb/Text/expected/hit_testing/empty-anonymous-container-overlapping.txt

<#text>
index: 4
<A id="target">
